requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.5.2
camel-converter==4.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
curl_cffi==0.11.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-redis==0.2.1
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
meilisearch==0.34.1
mistune==3.1.3
ml_dtypes==0.5.1
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.40.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
notebook==7.4.2
notebook_shim==0.2.4
numpy==2.2.6
openai==1.82.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==25.1.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.8
ply==3.11
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.9.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-ulid==3.0.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.3.0
redisvl==0.6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339_validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
Send2Trash==1.8.3
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack_data==0.6.3
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_utils==0.1.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.0
yfinance==0.2.61
zipp==3.21.0
zstandard==0.23.0
